refactor: separate dashboard routes
